# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BookTrack API",
        "description": "Documentação da API",
        "contact": {
            "email": "elessandrodev@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/autores": {
            "get": {
                "tags": [
                    "Autores"
                ],
                "summary": "Listar todos os autores",
                "operationId": "21f72ad76bb95b0c51e313b84b42a703",
                "responses": {
                    "200": {
                        "description": "Lista de autores"
                    }
                }
            },
            "post": {
                "tags": [
                    "Autores"
                ],
                "summary": "Cadastrar novo autor",
                "operationId": "c9f8d214d6250e0f085d91c214dead49",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "James Howlett"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Autor criado com sucesso"
                    }
                }
            }
        },
        "/api/autores/{id}": {
            "get": {
                "tags": [
                    "Autores"
                ],
                "summary": "Exibir autor específico",
                "operationId": "329de44608825d9d1a8b98e76bf81f49",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do autor"
                    },
                    "404": {
                        "description": "Autor não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Autores"
                ],
                "summary": "Atualizar autor",
                "operationId": "8e016da61bce1e8d2fe2b98fbdcc262e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "J. K. Rowling"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Autor atualizado com sucesso"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Autores"
                ],
                "summary": "Excluir autor",
                "operationId": "24f09c7c97f2c7e5bde2da5a7a81860f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Autor excluído com sucesso"
                    }
                }
            }
        },
        "/api/assuntos": {
            "get": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Listar todos os assuntos",
                "operationId": "04195ad80a665f44654634bf25e771e2",
                "responses": {
                    "200": {
                        "description": "Lista de assuntos"
                    }
                }
            },
            "post": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Cadastrar novo assunto",
                "operationId": "9dfe20b66469eee61ebc90266d1f5612",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Literatura"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Assunto criado com sucesso"
                    }
                }
            }
        },
        "/api/assuntos/{id}": {
            "get": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Exibir assunto específico",
                "operationId": "c15e5e63eca954996818836663535035",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do assunto"
                    },
                    "404": {
                        "description": "Assunto não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Atualizar assunto",
                "operationId": "ccaf7858efad1849b4e79dc0349cab8b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Literatura"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Assunto atualizado com sucesso"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Excluir assunto",
                "operationId": "4d06d55e9b72799ef0a03087fed4e892",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Assunto excluído com sucesso"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Assuntos",
            "description": "Gerenciamento de assuntos"
        },
        {
            "name": "Autores",
            "description": "Autores"
        }
    ]
}